feature: combos as keycodes
Problem:
LT() does not work as expected with combos, because combos are processed
before LT events. So if `PERMISSIVE_HOLD` is set and a combo is
triggered before the `TAPPING_TERM`, the combo is processed on the
current layer instead of the layer targetted by LT.

Fix:
We declare combos as keycodes, which allows LT to deal with the layer
change before the key event is processed.
A static array is used to know which combos are currently pressed, so
the right keycodes can be released in case of a layer change
between the combo press and release events.